Tidying a little
